requirements.txt
customtkinter>=5.2.2
yt-dlp>=2024.0.0
pillow>=10.0.0
packaging
